Fix(#48): 같은 클래스로 빈 등록 실패 해결
